requirements.txt
black==23.1.0
Django==4.1.6
django-autoslug==1.9.9
django-debug-toolbar==4.0.0
django-filter==23.2
django-formset==0.13.4
django-ranged-response==0.2.0
django-selectize==2.0
django-simple-captcha==0.5.17
django-templated-mail==1.1.1
django-uuslug==2.0.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.5
drf-yasg-json-api==0.9.1
geopy==2.3.0
psycopg2==2.9.2
psycopg2-binary==2.9.5
pytest==7.2.1
pytest-django==4.5.2
python-dotenv==1.0.0
